+++ src/main/java/com/monari/monariback/lesson/dto/request/CreateLessonRequest.java
@@ -30,7 +30,7 @@ import static com.monari.monariback.lesson.constant.LessonValidationConstants.TI
 import com.monari.monariback.common.enumerated.Region;
 import com.monari.monariback.common.enumerated.SchoolLevel;
 import com.monari.monariback.common.enumerated.Subject;
-import com.monari.monariback.lesson.entity.enurmurated.LessonStatus;
+import com.monari.monariback.lesson.entity.enurmerated.LessonStatus;
 import jakarta.validation.constraints.Future;
 import jakarta.validation.constraints.FutureOrPresent;
 import jakarta.validation.constraints.Min;
@@ -84,7 +84,7 @@ public record CreateLessonRequest(
 
     @NotNull(message = REGION_REQUIRED)
     Region region,
-    
+
     @NotNull(message = SCHOOL_LEVEL_REQUIRED)
     SchoolLevel schoolLevel,
