+++ src/main/java/com/monari/monariback/lesson/entity/enurmerated/LessonStatus.java
@@ -1,4 +1,4 @@
-package com.monari.monariback.lesson.entity.enurmurated;
+package com.monari.monariback.lesson.entity.enurmerated;
 
 public enum LessonStatus {
     ACTIVE,
